requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
networkx>=3.0
numpy>=1.24.0
scipy>=1.10.0

# Binary analysis
angr>=9.2.0
pyvex>=9.2.0

# Data processing
pandas>=2.0.0
scikit-learn>=1.3.0
pyyaml>=6.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.65.0

# Utilities
pyg-lib
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv

# Installation instructions:
# 1. Install PyTorch first (with CUDA if available):
#    pip install torch --index-url https://download.pytorch.org/whl/cu118
#
# 2. Install PyTorch Geometric:
#    pip install torch-geometric
#
# 3. Install other dependencies:
#    pip install -r requirements.txt
